pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pytorch-cosma"
version = "0.1.7"
description = "A deep learning framework built on top of PyTorch."
authors = [
    { name = "Ra'ad Costello-Mahmoud", email = "mahmoud.raad@yahoo.co.uk" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.9",
    "aiosignal==1.3.1",
    "async-timeout==5.0.1",
    "attrs==24.2.0",
    "blinker",
    "Bottleneck",
    "Brotli",
    "choreographer==0.99.6",
    "daal4py==2023.1.1",
    "dash",
    "dash-vtk==0.0.9",
    "dash_cytoscape==1.0.2",
    "filelock==3.16.1",
    "Flask",
    "Flask-Compress",
    "frozenlist==1.5.0",
    "fsspec==2024.9.0",
    "matplotlib==3.9.2",
    "ml_dtypes==0.5.0",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "networkx==3.4.1",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "opencv-python",
    "packaging==24.2",
    "pandas",
    "pillow==11.0.0",
    "ply==3.11",
    "protobuf==4.25.3",
    "psutil",
    "pycocotools",
    "pydantic",
    "pydantic_core",
    "Pygments",
    "pynndescent",
    "PyQt5==5.15.10",
    "PySocks",
    "python-dateutil",
    "pytz",
    "PyYAML",
    "pyzmq",
    "requests",
    "retrying",
    "scikit-learn",
    "scikit-learn-intelex",
    "scipy",
    "sip",
    "six",
    "tenacity",
    "threadpoolctl",
    "torch==2.5.1",
    "torch-geometric==2.6.1",
    "torchaudio==2.5.1",
    "torchvision==0.20.1",
    "tornado",
    "tqdm",
    "triton==3.1.0",
    "typing_extensions==4.12.2",
    "umap-learn",
    "unicodedata2",
    "urllib3",
    "vtk==9.4.0",
    "Werkzeug",
    "yarl==1.18.3",
    "zipp"
]

[project.urls]
homepage = "https://github.com/raadcostellomahmoud/pytorch-cosma"
repository = "https://github.com/raadcostellomahmoud/pytorch-cosma"


[tool.setuptools.packages]
find = { include = ["pytorch_cosma"] }  # Only include this package